% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbdenton.R
\name{denton.modelbased}
\alias{denton.modelbased}
\title{Title}
\usage{
denton.modelbased(
  series,
  indicator,
  differencing = 1,
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  conversion.obsposition = 1,
  outliers = NULL
)
}
\arguments{
\item{series}{}

\item{indicator}{}

\item{differencing}{Not implemented yet}

\item{conversion}{}

\item{conversion.obsposition}{}

\item{outliers}{}
}
\value{

}
\description{
Title
}
