% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\name{temporaldisaggregation}
\alias{temporaldisaggregation}
\title{Temporal disaggregation of a time series by regression models. Includes Chow-Lin, Fernandez, Litterman and some variants of those algorithms}
\usage{
temporaldisaggregation(
  series,
  constant = T,
  trend = F,
  indicators = NULL,
  model = c("Ar1", "Rw", "RwAr1"),
  freq = 4,
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  conversion.obsposition = 1,
  rho = 0,
  rho.fixed = F,
  rho.truncated = 0,
  zeroinitialization = F,
  diffuse.algorithm = c("SqrtDiffuse", "Diffuse", "Augmented"),
  diffuse.regressors = F
)
}
\arguments{
\item{series}{The time series that will be disaggregated}

\item{constant}{Constant term (T/F)}

\item{trend}{Linear trend (T/F)}

\item{indicators}{High-frequency indicator used in the temporal disaggregation}

\item{model}{Model of the error term (at the disaggregated level). "Ar1" = Chow-Lin, "Rw" = Fernandez, "RwAr1" = Litterman}

\item{freq}{Annual frequency of the disaggregated variable. Used if no indicator is provided}

\item{conversion}{Conversion mode (Usually "Sum" or "Average")}

\item{conversion.obsposition}{Only used with "UserDefined" mode. Position of the observed indicator in the aggregated periods (for instance 7th month of the year)}

\item{rho}{Only used with Ar1/RwAr1 models. (Initial) value of the parameter}

\item{rho.fixed}{Fixed rho (T/F, F by default)}

\item{rho.truncated}{Range for Rho evaluation (in [rho.truncated, 1[)}

\item{zeroinitialization}{The initial values of an auto-regressive model are fixed to 0 (T/F, F by default)}

\item{diffuse.algorithm}{Algorithm used for diffuse initialization. "SqrtDiffuse" by default}

\item{diffuse.regressors}{Indicates if the coefficients of the regression model are diffuse (T) or fixed unknown (F, default)}
}
\value{
An object of class "JD3TempDisagg"
}
\description{
Temporal disaggregation of a time series by regression models. Includes Chow-Lin, Fernandez, Litterman and some variants of those algorithms
}
\examples{
Y<-rjd3toolkit::aggregate(rjd3toolkit::retail$RetailSalesTotal, 1)
x<-rjd3toolkit::retail$FoodAndBeverageStores
td<-rjd3bench::temporaldisaggregation(Y, indicators=x)
y<-td$estimation$disagg
}
