% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\name{temporaldisaggregation2}
\alias{temporaldisaggregation2}
\title{Temporal disaggregation without indicator}
\usage{
temporaldisaggregation2(
  series,
  indicator,
  model = c("Ar1", "Rw"),
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  conversion.obsposition = 1,
  rho = 0,
  rho.fixed = F,
  rho.truncated = 0
)
}
\arguments{
\item{rho.truncated}{}
}
\value{

}
\description{
Temporal disaggregation without indicator
}
