% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempdisagg.R
\name{temporaldisaggregationI}
\alias{temporaldisaggregationI}
\title{Temporal disaggregation using the model: x(t) = a + b y(t), where x(t) is the indicator, 
y(t) is the unknown target series, with low-frequency constraints on y.}
\usage{
temporaldisaggregationI(
  series,
  indicator,
  model = c("Ar1", "Rw"),
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  conversion.obsposition = 1,
  rho = 0,
  rho.fixed = F,
  rho.truncated = 0
)
}
\arguments{
\item{series}{The time series that will be disaggregated}

\item{indicator}{High-frequency indicator used in the temporal disaggregation}

\item{model}{Model of the error term (at the disaggregated level). "Ar1" = Chow-Lin, "Rw" = Fernandez, "RwAr1" = Litterman}

\item{conversion}{Conversion mode (Usually "Sum" or "Average")}

\item{conversion.obsposition}{Only used with "UserDefined" mode. Position of the observed indicator in the aggregated periods (for instance 7th month of the year)}

\item{rho}{Only used with Ar1/RwAr1 models. (Initial) value of the parameter}

\item{rho.fixed}{Fixed rho (T/F, F by default)}

\item{rho.truncated}{Range for Rho evaluation (in [rho.truncated, 1[)}

\item{freq}{Annual frequency of the disaggregated variable. Used if no indicator is provided}
}
\value{
An object of class "JD3TempDisaggI"
}
\description{
Temporal disaggregation using the model: x(t) = a + b y(t), where x(t) is the indicator, 
y(t) is the unknown target series, with low-frequency constraints on y.
}
\examples{
Y<-rjd3toolkit::aggregate(rjd3toolkit::retail$RetailSalesTotal, 1)
x<-rjd3toolkit::retail$FoodAndBeverageStores
td<-rjd3bench::temporaldisaggregationI(Y, indicator=x)
y<-td$estimation$disagg
}
